+++ src/hypofuzz/__init__.py
@@ -1,4 +1,4 @@
 """Adaptive fuzzing for property-based tests using Hypothesis."""
 
-__version__ = "25.01.4"
+__version__ = "25.01.5"
 __all__: list = []